requirements.txt
squareup >= 5.0.0